+++ src/subramanian/Assignment_22.java
@@ -1,4 +1,4 @@
-package subramanium;
+package subramanian;
 
 public class Assignment_22 {
