+++ src/subramanian/CharFreq.java
@@ -1,4 +1,4 @@
-package subramanium;
+package subramanian;
 
 public class CharFreq {
